% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statar.R
\name{arc.miles}
\alias{arc.miles}
\title{Arc Miles}
\usage{
arc.miles(lat1, long1, lat2, long2)
}
\arguments{
\item{lat1}{the latitude of the first location}

\item{long1}{the longitude of the first location}

\item{lat2}{the latitude of the second location}

\item{long2}{the longitude of the second location}
}
\description{
Returns the arc miles between lat1, long1, and lat2, long2
}
\examples{
arc.miles(49.123, -119.33, 49.02, -118.12)
}
\keyword{arc}
\keyword{miles}
