% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statar.R
\name{as.rad}
\alias{as.rad}
\title{As Radians}
\usage{
as.rad(degrees)
}
\arguments{
\item{degrees}{the degrees for which you'd like radians}
}
\description{
Returns radians from degrees
}
\examples{
as.rad(365)
}
\keyword{degrees}
\keyword{rad}
\keyword{to}
