% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binscatter.R
\name{binscatter}
\alias{binscatter}
\alias{binscatter_}
\title{Experimental function to graph a dataset}
\usage{
binscatter(x, formula, w = NULL, verbose = FALSE, n = 20)

binscatter_(x, formula, w = NULL, verbose = FALSE, n = 20)
}
\arguments{
\item{x}{A data.frame}

\item{formula}{A formula.}

\item{w}{Wweights}

\item{n}{Number of quantiles to plot}

\item{A}{plot of the lhs of the formula over the first regressor, after controlling by variables in the formula.}
}
\description{
Experimental function to graph a dataset
}
\examples{
library(dplyr)
binscatter(iris, Sepal.Width ~ Sepal.Length)
binscatter(iris, Sepal.Width ~ Sepal.Length | Species, n = 10)
binscatter(group_by(iris, Species), Sepal.Width ~ Sepal.Length, n = 10)
}

