% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cluster}
\alias{cluster}
\title{Create quantile categories (corresponds to Stata command xtile)}
\usage{
cluster(x, nq = NULL, cutpoints = NULL)
}
\arguments{
\item{x}{A vector}

\item{nq}{Number of quantiles. Quantiles are computed as the inverse of the empirical distribution function}

\item{cutpoints}{Cutpoints to use when \code{nq} is not specified.  For instance \code{cutpoints = 0.4} creates two groups, one for observations equal or below 0.4, one for observations superior to 0.4}
}
\description{
Create quantile categories (corresponds to Stata command xtile)
}
\examples{
library(dplyr)
DT \%>\% group_by(id) \%>\% mutate(new = cluster(date, nq = 3))
DT \%>\% group_by(v1) \%>\% mutate(cluster(v2, nq = 3))
DT \%>\% group_by(v1) \%>\% mutate(cluster(v2, cutpoints = c(1e5,5e5) ))
}

