% Generated by roxygen2 (4.0.2): do not edit by hand
\name{col_order}
\alias{col_order}
\title{Move variables to the front (Stata command order)}
\usage{
col_order(.data, ..., inplace = FALSE)
}
\arguments{
\item{...}{Variables to include/exclude in s You can use same specifications as in select. If missing, defaults to all non-grouping variables.}

\item{inplace}{Change data.table in place or not}

\item{DT}{A tbl_dt or tbl_grouped_dt.}
}
\description{
Move variables to the front (Stata command order)
}
\examples{
library(data.table)
library(dplyr)
N <- 100; K <- 10
DT <- data.table(
  id = 1:N,
  v1 =  sample(5, N, TRUE),
  v2 =  sample(1e6, N, TRUE),
  v3 =  sample(round(runif(100, max = 100), 4), N, TRUE)
)
DT  \%>\% col_order(starts_with("v"), inplace = TRUE)
}

