% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\title{Count number of non missing observations}
\usage{
count(...)
}
\arguments{
\item{...}{A set of vectors}
}
\description{
Count number of non missing observations
}
\examples{
count(1:100, c(NA, 1:99))
}

