% Generated by roxygen2 (4.0.2): do not edit by hand
\name{count_combinations}
\alias{count_combinations}
\title{Find best string combinations that identify an id}
\usage{
count_combinations(name, by, n = 1)
}
\arguments{
\item{name}{a vector of characters}

\item{by}{a vector of identifiers}

\item{n}{number of words for combinations. Default to \code{1}.}
}
\value{
\code{tab_accross} returns a data.frame of four columns. The first is id, the second corresponds to unique combination of words in each element of \code{v} with length lower than \code{n} (sorted alphabetically),  the third is the count of these permutation within \code{id}, the fourth is the count of these permutation accross \code{i}. When the count accross group is 1 and the count within group is high, the element can be considered as an identifier of the group.
library(data.table)
id <- c(1, 1, 2, 2)
name <- c("coca cola company", "coca cola incorporated", "apple incorporated", "apple corp")
count_combinations(name, by = id)
count_combinations(name, by = id, n = 2)
}
\description{
Find best string combinations that identify an id
}

