% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set.R
\name{discard}
\alias{discard}
\alias{discard_}
\title{Create a new data.table by discarding certain columns}
\usage{
discard(x, ...)

discard_(x, vars)
}
\arguments{
\item{x}{a data.table}

\item{...}{Variables to keep. Default to all. See the \link[dplyr]{select} documentation.}

\item{vars}{Used to work around non-standard evaluation.}
}
\description{
Create a new data.table by discarding certain columns
}
\examples{
library(data.table)
DT <- data.table(
  id = c(1,2),
  v1 = c(1,1),
  v2 = c(2,1)
)
discard(DT, id, v2)
discard(DT, -id)
}

