% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drop}
\alias{drop}
\alias{drop_}
\title{Create new data.table by dropping certain columns}
\usage{
drop(x, ...)

drop_(x, vars)
}
\arguments{
\item{x}{a data.table}

\item{...}{Variables to keep. Default to all. See the \link[dplyr]{select} documentation.}

\item{vars}{Used to work around non-standard evaluation.}
}
\description{
Create new data.table by dropping certain columns
}
\examples{
library(data.table)
DT <- data.table(
  id = c(1,NA,2),
  v1 = c(1,1,2)
)
drop_if(DT, id==1)
}

