% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drop_if}
\alias{drop_if}
\alias{drop_if_}
\title{Create new data.table by dropping certain rows}
\usage{
drop_if(x, ..., by = NULL)

drop_if_(x, .dots, by)
}
\arguments{
\item{x}{a data.table}

\item{...}{Conditions. Rows where the condition evaluates to NA are not dropped. Therefore, \code{drop_if(dt, condition)} is not the same as \code{keep_if(x, !condition)} with}

\item{by}{groups in which the condition should be evaluated}

\item{.dots}{Used to work around non-standard evaluation.}
}
\description{
Create new data.table by dropping certain rows
}
\examples{
library(data.table)
DT <- data.table(
  id = c(1,NA,2),
  v1 = c(1,1,2)
)
drop_if(DT, id==1)
}

