% Generated by roxygen2 (4.0.2): do not edit by hand
\name{duplicates}
\alias{duplicates}
\alias{duplicates_}
\title{Returns a data.table with duplicated groups}
\usage{
duplicates(x, ..., by = NULL)

duplicates_(x, ..., .dots, by = NULL)
}
\arguments{
\item{x}{a data.table}

\item{...}{Variables to keep (beyond the grouped). Default to all variables. See the \link[dplyr]{select} documentation.}

\item{by}{Variable to group by. Default is the key, or everything is the data.table is not keyed.}
}
\value{
a data.table composed of groups that have duplicates. The first column is a new variable, named "N", that displays the number of duplicates
}
\description{
Returns a data.table with duplicated groups
}
\examples{
DT <- data.table(a = rep(1:2, each = 3), b = 1:6)
duplicates(DT, by = "a")
}

