% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edo}
\alias{edo}
\title{Set for Stata commands that don't modify datasets}
\usage{
edo(DT, fun, cols = names(DT), ..., i = NULL, by = NULL)
}
\arguments{
\item{DT}{A data.table.}

\item{cols}{A character vector of columns on which to apply the command.}

\item{...}{Options to pass to the datata command.
DT %>% edo(summarize, "v2")
DT %>% edo(sum, "v*", d = TRUE)}

\item{cmd}{One stata commandout of the following: summarize. Abbreviations are accepted.}
}
\description{
Set for Stata commands that don't modify datasets
}

