% Generated by roxygen2 (4.0.2): do not edit by hand
\name{elapsed}
\alias{as.Date.monthly}
\alias{as.Date.quarterly}
\alias{as.Date.weekly}
\alias{as.POSIXlt.monthly}
\alias{as.POSIXlt.quarterly}
\alias{as.POSIXlt.weekly}
\alias{as.data.frame.monthly}
\alias{as.data.frame.quarterly}
\alias{as.data.frame.weekly}
\alias{as.monthly}
\alias{as.quarterly}
\alias{as.weekly}
\alias{elapsed}
\alias{format.monthly}
\alias{format.quarterly}
\alias{format.weekly}
\alias{monthly,}
\alias{print.monthly}
\alias{print.quarterly}
\alias{print.weekly}
\alias{quarterly,}
\alias{weekly}
\title{Elapsed calendar periods}
\usage{
as.quarterly(date)

as.monthly(date)

as.weekly(date)

\method{as.POSIXlt}{quarterly}(date, ...)

\method{as.POSIXlt}{monthly}(date, ...)

\method{as.POSIXlt}{weekly}(date, ...)

\method{as.Date}{quarterly}(date, ...)

\method{as.Date}{monthly}(date, ...)

\method{as.Date}{weekly}(date, ...)

\method{format}{quarterly}(date, ...)

\method{format}{monthly}(date, ...)

\method{format}{weekly}(date, ...)

\method{as.data.frame}{monthly}(...)

\method{as.data.frame}{quarterly}(...)

\method{as.data.frame}{weekly}(...)

\method{print}{quarterly}(date)

\method{print}{monthly}(date)

\method{print}{weekly}(date)
}
\arguments{
\item{date}{Any vector that has a method for "}
}
\description{
Elapsed calendar periods
}
\details{
Weekly, monthly and quarterly dates are dates stored as the number of elapsed calendar periods since January 01 1970. This is helpful for two reasons. First, it allows to use methods defined on numeric on quarterly date (for instance lag wrt one period). Moreover, these classes are printed in a way that fits their frequency YYYYqQ, YYYYmMM, YYYYmWW.  Methods to convert from and to Dates or POSIXlt are provided.
}
\examples{
date <- mdy(c("04/03/1992", "01/04/1992", "03/15/1992"))
l <- data.frame(
   weekly = as.weekly(date),
   monthly = as.monthly(date),
   quarterly = as.quarterly(date)
   )
l
l+1
l \%>\% mutate_each(funs(week))
l \%>\% mutate_each(funs(lag(c(1,2,3), along_with = .)))
}

