% Generated by roxygen2 (4.0.2): do not edit by hand
\name{epanel}
\alias{epanel}
\title{Set of commands for panel data}
\usage{
epanel(DT, cols, time, cmd, gen = cmdc, inplace = FALSE)
}
\arguments{
\item{DT}{A data.table.}

\item{cols}{A character vector corresponding to the panel id}

\item{time}{A character vector corresponding to time}

\item{cmd}{Commmand to use. Can be of the form L3.varlue or fill}

\item{inplace}{A boolean that specifies whether the dataset should be modified in place or not (option is only available wheb creating lagged/forward variables)}
}
\description{
Set of commands for panel data
}
\examples{
library(data.table)
library(dplyr)
DT <- data.table(
 id = c(1, 1, 1, 1, 1, 2, 2),
 date = c(1992, 1989, 1991, 1990, 1994, 1992, 1991),
 value = c(4.1, 4.5, 3.3, 5.3, 3.0, 3.2, 5.2)
)
DT <- DT \%>\% epanel(cols = "id", t = "date", L1.value, inplace = FALSE)
DT \%>\% epanel(cols = "id", t = "date", L3.value, gen = "L3.value", inplace = TRUE)
DT <- DT \%>\% epanel(cols = "id", t = "date", fill)
}

