% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eset}
\alias{eset}
\title{Set for Stata commands that modify the dataset}
\usage{
eset(DT, fun, cols = names(DT), ..., i = NULL, by = NULL)
}
\arguments{
\item{DT}{A data.table.}

\item{cols}{A character vector of columns on which to apply the command.}

\item{...}{Options to pass to the datata command.
DT %>% eset(order, cols = "v*")
DT %>% eset(sort, c("v1", "v2"))
DT %>% eset(rename, "v1", "v11")
DT %>% eset(keep, -"id*")
DT %>% eset(keep, "v?")}

\item{cmd}{One stata commandout of the following: sort, order, rename, keep, drop. Abbreviations are accepted.}
}
\description{
Set for Stata commands that modify the dataset
}

