% Generated by roxygen2 (4.0.2): do not edit by hand
\name{evalm}
\alias{evalm}
\alias{quotem}
\title{String and expression interpolation}
\usage{
evalm(x, env = parent.frame(), inherits = FALSE)

quotem(x, env = parent.frame(), inherits = FALSE)
}
\arguments{
\item{x}{any syntactically valid R expression}

\item{env}{environment in which to evalute the expressions enclosed in \code{$}}

\item{inherits}{Default to FALSE}
}
\value{
\cite{quotem} captures the (unevaluated) expression given as an argument and returns it after replacing any expression starting with \code{$} by its evaluated expression.

\cite{evalm} is a wrapper for \cite{eval(quotem(exp))}
}
\description{
String and expression interpolation
}
\examples{
name="Bob"
height=72
weight=230
quotem("My record indicates you are $height inches tall and weight $weight pounds")
quotem("Your body mass index is $(round(703*weight/height^2,1))")
newvar <- "temp"
within <- "var"
quotem("my new variable is $(new$within)")
library(data.table)
N <- 100
DT <- data.table(
  id = sample(5, N, TRUE),
  v1 = sample(5, N, TRUE),
  v2 = sample(1e6, N, TRUE)
)
myvar <- "v1"
byvar <- "id"
quotem(DT[, list(`$newvar` = mean(`$myvar`)), by = `$byvar`])
evalm(DT[, list(`$newvar` = mean(`$myvar`)), by = `$byvar`])
}

