% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expand.data.table}
\alias{expand.data.table}
\title{Fill Absent observations}
\usage{
expand.data.table(.data, ..., type = c("within", "across"))
}
\arguments{
\item{.data}{A tbl_dt, grouped or not}

\item{...}{Variables to expand}
}
\description{
Fill Absent observations
}
\examples{
library(data.table)
library(dplyr)
DT <- data.table(
 id = c(1, 1, 1, 1, 1, 2, 2),
 date = c(1992, 1989, 1991, 1990, 1994, 1992, 1991),
 value = c(4.1, 4.5, 3.3, 5.3, 3.0, 3.2, 5.2)
)
DT <- DT \%>\% group_by(id) \%>\% expand(date)
DT <- DT \%>\% expand(date)
}

