% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/find_duplicates.R
\name{find_duplicates}
\alias{find_duplicates}
\alias{find_duplicates_}
\title{returns a data.table with duplicated rows}
\usage{
find_duplicates(x, ..., gen = "N")

find_duplicates_(x, vars, gen = "N")
}
\arguments{
\item{x}{a data.table}

\item{...}{Variable on which one should check for duplicates. Default to all variables}

\item{gen}{A character that specifies  the name of a new variable with the number of duplicates. Default to "N".}

\item{vars}{Used to work around non-standard evaluation.}
}
\value{
a data.table with groups that have duplicates.
}
\description{
returns a data.table with duplicated rows
}
\examples{
library(data.table)
DT <- data.table(a = rep(1:2, each = 3), b = 1:6)
find_duplicates(DT, a)
find_duplicates(DT, a, b)
}

