% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/floor_date.R
\name{floor_date}
\alias{floor_date}
\title{floor_date
Round date-times down.}
\usage{
floor_date(x, unit = c("second", "minute", "hour", "day", "week", "month",
  "quarter", "year"))
}
\arguments{
\item{x}{a vector of date-time objects}

\item{unit}{a character string specifying the time unit to be rounded to. Should be one of
"second","minute","hour","day", "week", "month", or "year."}
}
\value{
x with the appropriate units floored
}
\description{
floor_date takes a date-time object and rounds it down to the nearest integer
value of the specified time unit. Users can specify whether to round down to
the nearest second, minute, hour, day, week, month, quarter or year.
}

