% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fuzzy_join}
\alias{fuzzy_join}
\title{Join two data.tables together}
\usage{
fuzzy_join(x, y, exact = NULL, exact.or.NA = NULL, fuzzy = NULL,
  w = rep(1, length(fuzzy)), gen = "score", which = FALSE,
  method = "jw", p = 0.1)
}
\arguments{
\item{x}{The master data.table}

\item{y}{The using data.table}

\item{exact.or.NA}{Exact or NA}

\item{fuzzy}{Fuzzy columns on which to match}

\item{w}{Vector of weights}

\item{on}{Exact: exact characters}

\item{methods}{see stringdist}

\item{...}{argument to pass to stringdist}
}
\description{
Join two data.tables together
}

