% Generated by roxygen2 (4.0.2): do not edit by hand
\name{graph}
\alias{graph}
\alias{graph_}
\title{Experimental function to graph a dataset}
\usage{
graph(x, ..., along_with = NULL, by = NULL, w = NULL, reorder = TRUE,
  winsorize = TRUE, facet = FALSE, size = 1, verbose = FALSE)

graph_(x, ..., .dots, along_with = NULL, by = NULL, w = NULL, d = FALSE,
  reorder = TRUE, winsorize = winsorize, facet = FALSE, size = 1,
  verbose = FALSE)
}
\arguments{
\item{...}{Variables to include/exclude. Defaults to all non-grouping variables. See the \link[dplyr]{select} documentation.}

\item{along_with}{Replace x axis from percentiles by this variable.}

\item{by}{Groups within which variables should be ploted.}

\item{reorder}{Should the category with the most count be printed first?}

\item{winsorize}{Should numeric variables winsorized?}

\item{facet}{Should results graphed in different windows for each group?}

\item{size}{Point sizes when more than 1000 points by group}

\item{verbose}{Should warnings (regarding missing values, outliers, etc) be printed?}

\item{DT}{A tbl_dt or tbl_grouped_dt.}
}
\description{
Experimental function to graph a dataset
}
\examples{
N <- 10000
DT <- data.table(
  id = sample(c("id1","id2","id3"), N, TRUE),
  v1 = sample(c(1:5), N, TRUE),
  v2 = rnorm(N, sd = 20),
  v3 = sample(runif(100, max=100), N, TRUE)
)
DT[, v4 := v3 + rnorm(N, sd = 20)]
graph(DT)
graph(DT, by = id)
graph(DT, by = id, facet = TRUE)
graph(DT, v3, v4, along_with = v2)
graph(DT, v3, v4, along_with = v2, by = id)
}

