% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{Group multiple variable (similar to Stata group)}
\usage{
group(...)
}
\arguments{
\item{...}{vectors}
}
\value{
An integer vector representing groups
}
\description{
Group multiple variable (similar to Stata group)
}
\examples{
library(dplyr)
mutate(iris, g = group(Species))
mutate(iris, g = group(Species, floor(Sepal.Width)))
}

