% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{id}
\alias{id}
\title{Find string that identifies an id}
\format{\preformatted{ num [1:4] 1 1 2 2
}}
\usage{
id
}
\arguments{
\item{id}{identifier}

\item{v}{string associated to id}
}
\description{
\code{fuzzy_join} uses record linkage methods to match observations between two datasets where no perfect key fields exist.  For each row in x, \code{fuzzy_join} finds the closest row(s) in y. The distance is a weighted average of the string distances defined in \code{method} over multiple columns.
}
\keyword{datasets}

