% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is_na}
\alias{is_na}
\title{Returns boolean vector corresponding to rows with one missing observation in variable specified in "by"}
\usage{
is_na(x, by = seq_along(x))
}
\arguments{
\item{x}{a data.table}

\item{by}{a character or numeric vector for columns}
}
\description{
lead and lag are useful for comparing values for date offset by a constant
}
\examples{
library(data.table)
DT <- data.table( v1 = c(NA, 1), v2 = c(1, NA))
is_na(DT, by = "v1")
is_na(DT, by = "v2")
is_na(DT)
}

