% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statar.R
\name{isid}
\alias{isid}
\title{Is Variable Uniquely Identifying}
\usage{
isid(data)
}
\arguments{
\item{data}{the data for which you want to know if uniquely identifying}
}
\description{
whether the data is unique - combinations of variables too!
}
\examples{
data(mtcars) # load the cars dataset
isid(row.names(mtcars))
}
\keyword{identifier}
\keyword{unique}
