% Generated by roxygen2 (4.0.2): do not edit by hand
\name{keep_if}
\alias{keep_if}
\alias{keep_if_}
\title{Create new data.table by keeping only certain rows}
\usage{
keep_if(x, ..., by = NULL)

keep_if_(x, .dots, by)
}
\arguments{
\item{x}{a data.table}

\item{...}{Conditions}

\item{by}{groups in which the condition should be evaluated}

\item{.dots}{Used to work around non-standard evaluation.}
}
\description{
Create new data.table by keeping only certain rows
}
\examples{
library(data.table)
DT <- data.table(
  id = c(1,NA,2),
  v1 = c(1,1,2)
)
keep_if(DT, id==1)
}

