% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_narm.R
\name{n_narm}
\alias{n_narm}
\title{Count number of non missing observations}
\usage{
n_narm(...)
}
\arguments{
\item{...}{a sequence of vectors, matrices and data frames.}
}
\description{
Count number of non missing observations
}
\examples{
                         
n_narm(1:100, c(NA, 1:99))
}
