% Generated by roxygen2 (4.0.2): do not edit by hand
\name{n_unique.data.table}
\alias{n_unique.data.table}
\title{returns number of unique element in a data.table}
\usage{
\method{n_unique}{data.table}(x, by = key(x), ...)
}
\arguments{
\item{x}{a vector or a data.table}

\item{by}{Character vector. Variables that define groups. Default to key}

\item{...}{generics compatibility}
}
\value{
a data.table with groups that have duplicates.
}
\description{
returns number of unique element in a data.table
}

