% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statar.R
\name{odds}
\alias{odds}
\title{Odds Given Probability}
\usage{
odds(x)
}
\arguments{
\item{x}{the probability, between 0...1}
}
\description{
Returns the odds given the probability (0,1)
}
\examples{
odds(0.95)
}
\keyword{odds}
