% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statar.R
\name{OR}
\alias{OR}
\title{Odds Ratio}
\usage{
OR(y, x)
}
\arguments{
\item{y}{the controls}

\item{x}{the treatment}
}
\description{
Returns the odds ratio of y relative to x
}
\examples{
OR(5,8)
}
\keyword{OR}
\keyword{odds}
\keyword{ratio}
