% Generated by roxygen2 (4.0.2): do not edit by hand
\name{roll_lag}
\alias{roll_lag}
\alias{roll_lead}
\title{Apply rollling functions}
\usage{
roll_lag(x, FUN, n, along_with = NULL, order_by = NULL, ...)

roll_lead(x, FUN, n, along_with = NULL, order_by = NULL, ...)
}
\arguments{
\item{x}{a vector}

\item{FUN}{function to apply}

\item{n}{an integer specifying the rolling window}

\item{along_with}{use this variable to roll the function based on the  \code{[along_with - n, along_with]} rather than past \code{n} rows. NA are not accepted}

\item{order_by}{override the default ordering to use another vector}

\item{...}{options to pass to the function \code{FUN}}
}
\description{
Apply rollling functions
}
\examples{
along_with  = c(1, 2, 4, 7)
x = c(1, 1, 1, 1)
roll_lag(x, sum, n = 1, along_with = along_with)
roll_lag(x, sum, n = 2, along_with = along_with)
roll_lead(x, sum, n = 1, along_with = along_with)
roll_lead(x, sum, n = 2, along_with = along_with)
}

