% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/sample_corr.R, R/sample_cov.R
\name{sample_cov}
\alias{sample_cov}
\title{Sample correlation}
\usage{
sample_cov(x, y, w = rep(1L, length(x)))

sample_cov(x, y, w = rep(1L, length(x)))
}
\arguments{
\item{x}{A vector of values}

\item{y}{A vector of values}

\item{min}{If length of vector x once missing x and w and zero value of w are removed is strictly lower than min, the returned value is NA.}

\item{x}{A vector of values}

\item{y}{A vector of values}

\item{min}{If length of vector x once missing x and w and zero value of w are removed is strictly lower than min, the returned value is NA.}
}
\value{
Weighted covariance
}
\description{
Sample correlation

Sample covariance
}

