% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sample_mode}
\alias{sample_mode}
\title{Sample mode}
\usage{
sample_mode(x, na.rm = TRUE, ties.method = c("min", "max", "all"))
}
\arguments{
\item{x}{A vector of values}

\item{na.rm}{logical. Should NA ignored? Default to TRUE. If FALSE, NA are considered as a value per se.}

\item{ties}{A character among "all", "min" and "max" that changes behavior in case of multiple modes. Default to "min".}
}
\value{
Returns one mode of the vector
}
\description{
Sample mode
}
\examples{
sample_mode(c(1, 2, 2))
sample_mode(c(1, 2), ties = "min")
sample_mode(c(1, 2), ties = "max")
sample_mode(c(1, 2), ties = "all")
sample_mode(c(NA, NA, 1))
sample_mode(c(NA, NA, 1), na.rm = FALSE)
}

