% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/sample_sum.R
\name{sample_sum}
\alias{sample_sum}
\title{sum}
\usage{
sample_sum(x, w = rep(1L, length(x)), type = "aweight", min = 0)
}
\arguments{
\item{x}{A vector of values}

\item{w}{vector of weights.}

\item{type}{type of weights. Can be aweight (then weight are transformed so that sum to length of non missinx x)}

\item{min}{If length of vector x once missing x and w and zero value of w are removed is strictly lower than min, the returned value is NA.}

\item{na.rm}{logical. Should NA ignored? Default to TRUE. If FALSE, NA are considered as a value per se.}
}
\value{
When w / min are not specified,  sample_sum(, min = 0) is the same as sum(., na.rm = TRUE) while sample_sum(, min = length(x)) is the same as sum(.).
}
\description{
sum
}

