% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setcols}
\alias{setcols}
\title{keep columns}
\usage{
setcols(.data, cols)
}
\arguments{
\item{.data}{a data.table}

\item{cols}{a character vector of variable names to keep}
}
\description{
keep columns
}
\examples{
DT <- data.table(
 id    = c(1, 1, 1, 1, 1, 2, 2),
 date  = c(1992, 1989, 1991, 1993, 1994, 1992, 1991),
 value = c(NA, NA, 3, 5.3, 3.0, 3.2, 5.2)
)
setcols(DT, c("id", "date"))
}

