% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set.R
\name{setdiscard}
\alias{setdiscard}
\alias{setdiscard_}
\title{discard certain columns in place}
\usage{
setdiscard(x, ...)

setdiscard_(x, vars)
}
\arguments{
\item{x}{a data.table}

\item{...}{Variables to keep. Default to all. See the \link[dplyr]{select} documentation.}

\item{vars}{Used to work around non-standard evaluation.}
}
\description{
discard certain columns in place
}
\examples{
library(data.table)
DT <- data.table(
  id = c(1,2),
  v1 = c(1,1),
  v2 = c(2,1)
)
setdiscard(DT, id)
}

