% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setdrop}
\alias{setdrop}
\alias{setdrop_}
\title{Drop certain columns in place}
\usage{
setdrop(x, ...)

setdrop_(x, vars)
}
\arguments{
\item{x}{a data.table}

\item{...}{Variables to keep. Default to all. See the \link[dplyr]{select} documentation.}

\item{vars}{Used to work around non-standard evaluation.}
}
\description{
Drop certain columns in place
}
\examples{
library(data.table)
DT <- data.table(
  id = c(1,2),
  v1 = c(1,1),
  v2 = c(2,1)
)
setdrop(DT, id)
}

