% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setmutate}
\alias{setmutate}
\title{Version of mutate that (i) acts in place (ii) allows by, i condition}
\usage{
setmutate(x, ..., i = NULL, by = NULL)
}
\arguments{
\item{x}{a data.table}

\item{...}{Variables to include. Defaults to all non-grouping variables.}

\item{i}{a condition that restricts the set of row on which to apply the function}

\item{by}{Groups within which the function should be applied}

\item{replace}{If replace = TRUE, variables are replaced.}
}
\description{
Version of mutate that (i) acts in place (ii) allows by, i condition
}
\examples{
library(data.table)
N <- 100
DT <- data.table(
  id = sample(5, N, TRUE),
  v1 = sample(5, N, TRUE),
  v2 = sample(1e6, N, TRUE)
)
setmutate(DT, mean(v2), by = v1)
setmutate_(DT, setNames("mean(v2)", "v3"), by = "v1")
setmutate_(DT, setNames("mean(v2)", "v3"), i = "id==1", by = "v1")
}

