% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setpanel}
\alias{setpanel}
\alias{setpanel_}
\title{Check panel data}
\usage{
setpanel(x, ...)

setpanel_(x, vars)
}
\arguments{
\item{x}{a vector or matrix}

\item{...}{id variables. Last is time variable}

\item{vars}{Used to work around non standard evaluation}
}
\value{
The function checks no data variable is missing, that there are no duplicates for (id, along_with), and if both these conditions are satisfied, resort the data.table in place along the id and time variables.
}
\description{
Check panel data
}

