% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setroll}
\alias{setroll}
\alias{setroll_}
\title{Version of mutate_each that applies rollling functions}
\usage{
setroll(x, funs, ..., n, along_with, by = NULL, i = NULL)

setroll_(x, funs, vars, n, along_with, i = NULL, by = NULL)
}
\arguments{
\item{x}{a data.table}

\item{funs}{List of function calls, generated by funs, or a character vector of function names.}

\item{...}{Variables on which the functions should be applied. Defaults to all non-grouping variables. See the \link[dplyr]{select} documentation.}

\item{n}{an integer. Window is all the observations in \code{[along_with - n, along_with]}}

\item{along_with}{Numeric variable specifying dates. See the \link[dplyr]{select} documentation.}

\item{by}{Variables by which to group. Default to keys (or to keys minus last if along_with is unspecified). ee the \link[dplyr]{select} documentation.}
}
\description{
Version of mutate_each that applies rollling functions
}
\examples{
DT <- data.table(
 date =  c(1,2,4,7),
 value =  c(1,1,1,1)
)
DT1 <- copy(DT)
setroll(DT, funs(count = sum), value, n = 2,  along_with = date)
setroll(DT1, funs(count = sum), value, n = -2,  along_with = date)
}

