% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sum_up}
\alias{sum_up}
\title{Gives summary statistics (Stata command summarize)}
\usage{
sum_up(.data, ..., d = FALSE)
}
\arguments{
\item{...}{Variables to include/exclude in s You can use same specifications as in select. If missing, defaults to all non-grouping variables.}

\item{d}{Detail is true or not}

\item{DT}{A tbl_dt or tbl_grouped_dt.}
}
\description{
Gives summary statistics (Stata command summarize)
}
\examples{
library(data.table)
library(dplyr)
N <- 100; K <- 10
DT <- data.table(
  id = 1:N,
  v1 =  sample(5, N, TRUE),
  v2 =  sample(1e6, N, TRUE),
  v3 =  sample(round(runif(100, max = 100), 4), N, TRUE)
)
DT  \%>\% sum_up
DT  \%>\% sum_up(v3, d=T)
DT  \%>\% filter(v1==1) \%>\% sum_up(starts_with("v"))
}

