% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statar.R
\name{td}
\alias{td}
\title{Stata-like Dates}
\usage{
td(datein)
}
\arguments{
\item{datein}{a string of the date, formatted "01Jan1960"}
}
\description{
Stata-like date entry
}
\examples{
td("26Aug1984")
}
\keyword{date}
