% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tempname}
\alias{tempname}
\title{Create unique names within a list, a data.frame, or an environment}
\usage{
tempname(where = globalenv(), prefix = "temp", inherits = TRUE)
}
\arguments{
\item{where}{A chracter vector, list or an environment}

\item{prefix}{A character vector that specifies prefix for new name}

\item{inherits}{Should the name unique also in the enclosing frames of the environment?}
}
\description{
Create unique names within a list, a data.frame, or an environment
}
\examples{
tempname(c("temp","temp1"))
temp <- tempname(globalenv())
tempname(globalenv())
tempname(data.frame(temp = 1))
}

