% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tlead-tlag.R
\name{tlead-tlag}
\alias{tlag}
\alias{tlead}
\alias{tlead-tlag}
\title{lead and lag with respect to a time variable}
\usage{
tlead(x, n = 1L, along_with, default = NA)

tlag(x, n = 1L, along_with, default = NA)
}
\arguments{
\item{x}{a vector of values}

\item{n}{a positive integer of length 1, giving the number of positions to lead or lag by. When the package lubridate is loaded, it can be a period when using with along_with (see the lubridate function minutes, hours, days, weeks, months and years)}

\item{along_with}{time variable}

\item{default}{value used for non-existant rows. Defaults to \code{NA}.}
}
\description{
lead and lag with respect to a time variable
}
\examples{
year <- c(1989, 1991, 1992)
value <- c(4.1, 4.5, 3.3)
tlag(value, 1, along_with = year) #  returns value in year - 1
tlead(value, 1, along_with = year)
library(lubridate)
date <- mdy(c("01/04/1992", "03/15/1992", "04/03/1992"))
value <- c(4.1, 4.5, 3.3)
datem <- as.monthly(date)
tlag(value, along_with = datem)
}

