% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/sample_sd.R
\name{weighted_sd}
\alias{weighted_sd}
\title{Weighted sd}
\usage{
weighted_sd(x, w = rep(1L, length(x)), type = "aweight", min = 0)
}
\arguments{
\item{x}{A vector of values}

\item{w}{vector of weights.}

\item{type}{type of weights. Can be aweight (then weight are transformed so that sum to length of non missinx x)}

\item{min}{If length of vector x once missing x and w and zero value of w are removed is strictly lower than min, the returned value is NA.}
}
\value{
When w / min are not specified,  weighted_sd(, min = 0) is the same as sd(., na.rm = TRUE) while weighted_sd(, min = length(x)) is the same as sd(.).
}
\description{
Weighted sd
}

