% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wquantile.R
\name{wquantile}
\alias{wquantile}
\title{Weighted quantile of type 2 (similar to Stata _pctile)}
\usage{
wquantile(x, probs = c(0.25, 0.5, 0.75), w = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{A vector}

\item{w}{A weight vector}

\item{na.rm}{Should missing values be returned?}

\item{cutpoints}{Cutpoints to use when \code{nq} is not specified.  For instance \code{cutpoints = 0.4} creates two groups, one for observations equal or below 0.4, one for observations superior to 0.4.}
}
\description{
Weighted quantile of type 2 (similar to Stata _pctile)
}

