% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xtile}
\alias{xtile}
\title{Create variable containing quantiles (Stata command xtile)}
\usage{
xtile(x, nq = NULL, cutpoints = NULL)
}
\arguments{
\item{x}{A vector}

\item{nq}{Number of quantiles. Quantiles are computed as the inverse of empirical distribution function}

\item{cutpoints}{Cutpoints to use if nq is not specified.  For instance 0.4 creates two groups, one for observations equal or below 0.4, one for observations superior to 0.4}
}
\description{
Create variable containing quantiles (Stata command xtile)
}
\examples{
library(dplyr)
DT <- DT \%>\% group_by(id) \%>\% mutate(new = xtile(date, nq = 3))
DT \%>\% group_by(v1) \%>\% mutate(xtile(v2, nq = 3))
DT \%>\% group_by(v1) \%>\% mutate(xtile(v2, cutpoints = c(1e5,5e5) ))
}

