% Generated by roxygen2 (4.0.2): do not edit by hand
\name{demean}
\alias{demean}
\title{Demean a set of variables in a data.table}
\usage{
demean(x, ..., fe, suffix = "_demean", inplace = FALSE)
}
\arguments{
\item{x}{A data.table}

\item{...}{variables to demean}

\item{fe}{List of groups to use}
}
\value{
data.table with demeaned variables (NA if one of the variable is missing), with suffix "_demean"
}
\description{
Demean a set of variables in a data.table
}
\details{
This function calls felm::demeanlist after checking for missing values and converting group variables into factors
}
\examples{
N=1e3; K=100
set.seed(1)
DT <- data.table(
  id = sample(5, N, TRUE),
  v1 =  sample(5, N, TRUE),
  v2 =  sample(1e6, N, TRUE),
  v3 =  sample(round(runif(100,max=100),4), N, TRUE)
)
demean(DT, list(v2,v3), fe = list(id,v1))
}

