% Generated by roxygen2 (4.0.2): do not edit by hand
\name{duplicates}
\alias{duplicates}
\title{returns groups with duplicates}
\usage{
duplicates(x, by = key(x))
}
\arguments{
\item{x}{a data.table}

\item{by}{a character vectors. Default is key of data.table}
}
\value{
a data.table composed of groups that have duplicates. The first column is a new variable, named "N", that displays the number of duplicates
}
\description{
returns groups with duplicates
}
\examples{
DT <- data.table(a = rep(1:2, each = 3), b=1:6)
duplicates(DT, by = "a")
}

