% Generated by roxygen2 (4.0.2): do not edit by hand
\name{duplicates}
\alias{duplicates}
\alias{duplicates_}
\title{returns a data.table with duplicated rows}
\usage{
duplicates(x, ..., gen = "N")

duplicates_(x, vars, gen = "N")
}
\arguments{
\item{x}{a data.table}

\item{...}{Variable on which one should check for duplicates. Default to all variables}

\item{gen}{A character that specifies  the name of a new variable with the number of duplicates. Default to "N".}

\item{vars}{Used to work around non-standard evaluation.}
}
\value{
a data.table with groups that have duplicates.
}
\description{
returns a data.table with duplicated rows
}
\examples{
library(data.table)
DT <- data.table(a = rep(1:2, each = 3), b = 1:6)
duplicates(DT, a)
duplicates(DT, a, b)
}

