% Generated by roxygen2 (4.0.2): do not edit by hand
\name{duplicates}
\alias{duplicates}
\alias{duplicates_}
\title{returns duplicated rows}
\usage{
duplicates(x, ..., by = NULL)

duplicates_(x, ..., .dots, by = NULL)
}
\arguments{
\item{x}{a data.table}

\item{...}{Variables to keep (beyond the by variable). Default to all variables. See the \link[dplyr]{select} documentation.}

\item{by}{Variable to group by. Default is the key, or everything is the data.table is not keyed.}
}
\value{
a data.table with rows in groups that have duplicates. The first column is a new variable, called "N" (N-i if N already exists), that countains the number of duplicates
}
\description{
returns duplicated rows
}
\examples{
DT <- data.table(a = rep(1:2, each = 3), b = 1:6)
duplicates(DT, by = "a")
}

