% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edo}
\alias{edo}
\title{Set of Stata commands that don't modify datasets}
\usage{
edo(DT, cmd, cols = names(DT), ..., i = TRUE, by = NULL)
}
\arguments{
\item{DT}{A data.table.}

\item{cmd}{One stata commandout of the following: summarize. Abbreviations are accepted.}

\item{cols}{A character vector of columns on which to apply the command.}

\item{...}{Options to pass to the stata command.}

\item{i}{Apply commands on certain rows}

\item{by}{Apply commands within groups}
}
\description{
Set of Stata commands that don't modify datasets
}
\examples{
library(data.table)
library(dplyr)
N <- 100; K <- 10
DT <- data.table(
  id = 1:N,
  v1 =  sample(5, N, TRUE),
  v2 =  sample(1e6, N, TRUE),
  v3 =  sample(round(runif(100, max = 100), 4), N, TRUE)
)
DT \%>\% edo(summarize, "v2")
DT \%>\% edo(sum, "v*", d = TRUE)
DT \%>\% edo(sum, "v*", i = v1 == 1)
DT \%>\% edo(sum, "v*", d = TRUE, by = "v1")
}

