% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edo}
\alias{edo}
\title{Set of commands that don't replace data.table in place}
\usage{
edo(DT, fun, cols = names(DT), ..., i = NULL, by = NULL, d = FALSE)
}
\arguments{
\item{DT}{A data.table.}

\item{cols}{A character vector of columns on which to apply the command.}

\item{cmd}{One stata commandout of the following: summarize. Abbreviations are accepted.}
}
\description{
Set of commands that don't replace data.table in place
}

