% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edo}
\alias{edo}
\title{Set of Stata commands}
\usage{
edo(DT, fun, cols = names(DT), ..., i = NULL, by = NULL)
}
\arguments{
\item{DT}{A data.table.}

\item{cols}{A character vector of columns on which to apply the command.}

\item{...}{Options to pass to the datata command.
DT %>% edo(order, cols = "v*")
DT %>% edo(sort, c("v1", "v2"))
DT %>% edo(rename, "v1", "v11")
DT %>% edo(keep, -"id*")
DT %>% edo(keep, "v?")
DT %>% edo(summarize, "v2")
DT %>% edo(sum, "v*", d = TRUE)}

\item{cmd}{One stata commandout of the following: sort, order, rename, keep, drop, summarize. Abbreviations are accepted.}
}
\description{
Set of Stata commands
}

