% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ejoin}
\alias{ejoin}
\title{Join two data.tables together (similar to Stata command joinby)}
\usage{
ejoin(DTm, DTu, type, keep = c(c("master", "matched", "using")),
  gen = "_merge")
}
\arguments{
\item{DTm}{A data.table.}

\item{DTu}{A data.table.}

\item{type}{1:1 one to one merge, m:1 many to one merge, 1:m one to many, m:m many to many (all pairwise combinations)}

\item{keep}{A character vector that specifies rows to keep}

\item{gen}{A character that specifies name of new variable to mark result, or the boolean FALSE if no such variable is created. The variable equals 1 for rows in master only, 2 for rows in using only, 3 for matched rows.}
}
\value{
A data.table that joins rows in master and using datasets.
}
\description{
Join two data.tables together (similar to Stata command joinby)
}
\examples{
ejoin(DTm, DTu, type = 1:1, keep = c("master", "matched", "using"), gen = "_merge")
ejoin(DTm, DTu, m:1, keep = "matched")
ejoin(DTm, DTu, m:m, keep = c("master", "matched"), gen = FALSE)
}

