% Generated by roxygen2 (4.0.2): do not edit by hand
\name{elapsed}
\alias{as.monthly}
\alias{as.quarterly}
\alias{as.weekly}
\alias{elapsed}
\alias{monthly,}
\alias{quarterly,}
\alias{weekly}
\title{Elapsed dates (weekly, monthly, quarterly)}
\usage{
as.quarterly(date)

as.monthly(date)

as.weekly(date)
}
\description{
Elapsed dates (weekly, monthly, quarterly)
}
\details{
Weekly, monthly and quarterly dates are stored as the number of elapsed calendar periods since 01/01/1970. This allows to use arithmetic on dates other than daily. Moreover, these dates are printed in a way that fits their frequency  (\code{yyyy}q\code{q}, \code{yyyy}m\code{MM}, \code{yyyy}m\code{WW}).

Methods to convert from and to Dates or POSIXlt are provided. In particular, you may use lubridate \code{\link{week}} \code{\link{month}} and \code{\link{year}} to extract informations from elapsed dates.
}
\examples{
date <- mdy(c("04/03/1992", "01/04/1992", "03/15/1992"))
datem <-  as.monthly(date)
# Print depends on period
datem
as.character(datem)
# Internally stocked as inter, which allows things like:
datem+1
lag(c(1, 2, 3), n = 1, along_with = datem)
# Behaves as a date for lubridate operation
week(datem)
# Convert to usual date format if needed
as.POSIXlt(datem)
as.POSIXct(datem)
as.Date(datem)
}

