% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eset}
\alias{eset}
\title{Set of commands that replace data.table in place}
\usage{
eset(DT, cmd, cols = names(DT), ..., i = NULL, by = NULL, d = FALSE)
}
\arguments{
\item{DT}{A data.table.}

\item{cmd}{One stata commandout of the following: sort, order, rename, keep, drop. Abbreviations are accepted.}

\item{cols}{A character vector of columns on which to apply the command.}
}
\description{
Set of commands that replace data.table in place
}

