% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eset}
\alias{eset}
\title{Set of Stata commands that modify datasets}
\usage{
eset(DT, cmd, cols = names(DT), ...)
}
\arguments{
\item{DT}{A data.table.}

\item{cmd}{One stata commandout of the following: sort, order, rename, keep, drop. Abbreviations are accepted.}

\item{cols}{A character vector of columns on which to apply the command.}

\item{...}{Options to pass to the datata command.}
}
\description{
Set of Stata commands that modify datasets
}
\examples{
library(data.table)
library(dplyr)
N <- 100; K <- 10
DT <- data.table(
  id = 1:N,
  v1 =  sample(5, N, TRUE),
  v2 =  sample(1e6, N, TRUE),
  v3 =  sample(round(runif(100, max = 100), 4), N, TRUE)
)
DT \%>\% eset(order, cols = "v*")
DT \%>\% eset(sort, c("v1", "v2"))
DT \%>\% eset(rename, "v1", "v11")
DT \%>\% eset(keep, -"id*")
DT \%>\% eset(keep, "v?")
}

