% Generated by roxygen2 (4.0.2): do not edit by hand
\name{evalm}
\alias{evalm}
\alias{substitutem}
\alias{substitutem_}
\title{Substituting expressions enclosed in m() by the symbol of the character they refer to.}
\usage{
evalm(x, env = parent.frame(), inherits = FALSE)

substitutem_(x, env = parent.frame(), inherits = FALSE)

substitutem(x, env = parent.frame(), inherits = FALSE)
}
\arguments{
\item{x}{any syntactically valid R expression}

\item{env}{environment in which to evalute the expression enclosed in m()}

\item{inherits}{Default to FALSE}
}
\value{
evalm evaluates the expression given as an argument, after replacing any expression enclosed in m() by its evaluated result in env.
}
\description{
Substituting expressions enclosed in m() by the symbol of the character they refer to.
}
\examples{
name="Bob"
height=72
weight=230
evalm("My record indicates you are m(height) inches tall and weigh m(weight) pounds")
evalm("Your body mass index is m(round(703*weight/height^2,1))")
newvar <- "temp"
within <- "var"
evalm("my new variable is m(newm(within))")
library(data.table)
N <- 100
DT <- data.table(
  id = sample(5, N, TRUE),
  v1 = sample(5, N, TRUE),
  v2 = sample(1e6, N, TRUE)
)
myvar <- "v1"
byvar <- "id"
evalm(DT[, list(`m(newvar)` = mean(m(myvar))), by = m(byvar)])
}

