% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expand}
\alias{expand}
\alias{expand_}
\title{Fill in gaps in variable}
\usage{
expand(.data, ..., along_with, type = c("within", "across"))

expand_(.data, ..., along_with, .dots, type = c("within", "across"))
}
\arguments{
\item{.data}{A tbl_dt, grouped or not}

\item{...}{Variables to keep}

\item{along_with}{integer variable to expand}

\item{type}{"within"  means that rows are expanded with respect to min and max of \code{...} within groups (default) while "default" means that dates are expanded with respect to min and max of \code{...} across groups.}
}
\description{
Fill in gaps in variable
}
\examples{
library(data.table)
library(dplyr)
DT <- data.table(
 id = c(1, 1, 1, 1, 1, 2, 2),
 date = c(1992, 1989, 1991, 1990, 1994, 1992, 1991),
 value = c(4.1, 4.5, 3.3, 5.3, 3.0, 3.2, 5.2)
)
DT \%>\% group_by(id) \%>\% expand(value, along_with = date)
DT \%>\% group_by(id) \%>\% expand(value, along_with = date, type == "across")
}

