% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicates.R
\name{find_duplicates}
\alias{find_duplicates}
\alias{find_duplicates_}
\title{returns a data.frame with duplicated rows}
\usage{
find_duplicates(x, ...)

find_duplicates_(x, vars)
}
\arguments{
\item{x}{a data.frame}

\item{...}{Variable on which one should check for duplicates. Default to all variables}

\item{vars}{Used to work around non-standard evaluation.}
}
\value{
a data.frame with groups that have duplicates.
}
\description{
returns a data.frame with duplicated rows
}
\examples{
df <- data.frame(a = rep(1:2, each = 3), b = 1:6)
find_duplicates(df, a)
}
