% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fmode}
\alias{fmode}
\alias{sample_mode}
\title{Statistical mode}
\usage{
fmode(x, na.rm = TRUE, ties.method = c("min", "max", "all"))

sample_mode(x, na.rm = TRUE, ties.method = "min")
}
\arguments{
\item{x}{A vector of values}

\item{na.rm}{logical. Should NA ignored? Default to TRUE. If FALSE, NA are considered as a value per se.}

\item{ties}{A character among "all", "min" and "max"}
}
\value{
Returns one mode of the vector
}
\description{
Statistical mode
}
\examples{
fmode(c(1, 2, 2))
fmode(c(1, 2), ties = "min")
fmode(c(1, 2), ties = "max")
fmode(c(1, 2), ties = "all")
fmode(c(NA, NA, 1))
fmode(c(NA, NA, 1), na.rm = FALSE)
}

