% Generated by roxygen2 (4.0.2): do not edit by hand
\name{graph}
\alias{graph}
\alias{graph_}
\title{Experimental function to graph a dataset}
\usage{
graph(x, ..., along_with = NULL, by = NULL, w = NULL, reorder = TRUE,
  winsorize = TRUE, facet = FALSE, size = 1, verbose = FALSE,
  type = if (is.null(along_with)) "density" else "lm")

graph_(x, ..., .dots, along_with = NULL, by = NULL, w = NULL,
  reorder = TRUE, winsorize = TRUE, facet = FALSE, size = 1,
  verbose = FALSE, type = if (is.null(along_with)) "density" else "lm")
}
\arguments{
\item{x}{A data.table.}

\item{...}{Variables to include. Defaults to all non-grouping variables. See the \link[dplyr]{select} documentation.}

\item{along_with}{When "type" is "line", "lm" or "loeless", replace x axis by this variable (ie estimate regression models instead of density).}

\item{by}{Groups within which variables should be ploted.}

\item{reorder}{Should the category with the most count be printed first?}

\item{winsorize}{Should variables winsorized?}

\item{facet}{Should different groups graphed in different windows?}

\item{verbose}{Should warnings (regarding missing values, outliers, etc) be printed?}

\item{type}{type of graph among "density", "boxplot", "line", "lm", "loeless"}
}
\description{
Experimental function to graph a dataset
}
\examples{
N <- 10000
DT <- data.table(
  id = sample(c("id1","id2","id3"), N, TRUE),
  v1 = sample(c(1:5), N, TRUE),
  v2 = rnorm(N, sd = 20),
  v3 = sample(runif(100, max=100), N, TRUE)
)
DT[, v4 := v3 + rnorm(N, sd = 20)]
graph(DT)
graph(DT, by = id)
graph(DT, by = id, facet = TRUE)
graph(DT, by = id, type = "boxplot")
graph(DT, v3, v4, along_with = v2)
graph(DT, v3, v4, along_with = v2, by = id, type = "loeless")
}

