% Generated by roxygen2 (4.0.2): do not edit by hand
\name{graph}
\alias{graph}
\alias{graph_}
\title{Experimental function to graph a dataset}
\usage{
graph(x, ..., by = NULL, reorder = TRUE, winsorize = TRUE)

graph_(x, ..., .dots, by = NULL, d = FALSE, reorder = TRUE,
  winsorize = winsorize)
}
\arguments{
\item{...}{Variables to include/exclude. Defaults to all non-grouping variables. See the \link[dplyr]{select} documentation.}

\item{by}{Groups within which variables should be ploted.}

\item{reorder}{Should the value of strings with the most count be printed first?}

\item{winsorize}{Should numeric variables winsorized?}

\item{DT}{A tbl_dt or tbl_grouped_dt.}
}
\description{
Experimental function to graph a dataset
}
\examples{
N <- 100; K <- 10
DT <- data.table(
  v1 = sample(5, N, TRUE),
  v2 =  sample(round(runif(100,max=100),4), N, TRUE)
)
graph(DT)
graph(DT, v2, by = v1)
}

