% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{Create id for each group by order of appearance}
\usage{
group(x)
}
\arguments{
\item{x}{A vector, a list of vector, or a data.frame}
}
\description{
Create id for each group by order of appearance
}
\examples{
group(list(c(1, 2, 2, 2), c(4, 4, 5, 4)))
}

