% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{Group multiple variable (similar to Stata group)}
\usage{
group(..., na.rm = FALSE)
}
\arguments{
\item{...}{vectors}

\item{na.rm}{Should groups where some variable is NA return NA? Default to FALSE}
}
\value{
An integer vector representing groups
}
\description{
Group multiple variable (similar to Stata group)
}
\examples{
library(dplyr)
mutate(iris, g = group(Species))
mutate(iris, g = group(Species, floor(Sepal.Width)), na.rm = TRUE)
}

