% Generated by roxygen2 (4.0.2): do not edit by hand
\name{join}
\alias{join}
\title{Join two data.tables together}
\usage{
join(x, y, on = intersect(names(x), names(y)), type = "outer",
  gen = FALSE, check = c("m:m", "m:1", "1:m", "1:1"))
}
\arguments{
\item{x}{The master data.table}

\item{y}{The using data.table}

\item{on}{Character vectors specifying variables to match on. Default to common names between x and y.}

\item{type}{The type of (SQL) join among "outer" (default), "left", "right", "inner", "semi", "anti" and "cross".}

\item{gen}{Name of new variable to mark result, or the boolean FALSE (default) if no such variable should be created. The variable equals 1 for rows in master only, 2 for rows in using only, 3 for matched rows.}

\item{check}{A character checking for the presence of duplicates. Specifying "1:m" (resp "m:1") checks that joined variables uniquely identify observations in x (resp y).}
}
\value{
A data.table that joins rows in master and using datases. In order to avoid duplicates, identical variable names not joined are renamed with .x and .y suffixes. The order of data.tables x and y is not conserved, unless they are keyed.
}
\description{
Join two data.tables together
}
\examples{
join(x, y, on = intersect(names(x),names(y)), type = "outer", gen = FALSE, check = "m:m")
}

