% Generated by roxygen2 (4.0.2): do not edit by hand
\name{keep}
\alias{keep}
\alias{keep_}
\title{Create new data.table by keeping only certain columns or rows}
\usage{
keep(x, ..., i = NULL, by = NULL)

keep_(x, ..., .dots, i = NULL, by = NULL)
}
\arguments{
\item{x}{a data.table}

\item{...}{Variables to keep. Default to all. See the \link[dplyr]{select} documentation.}

\item{i}{Condition within groups}

\item{by}{Groups by which condition should be evaluated}

\item{.dots}{Used to work around non-standard evaluation.}
}
\description{
Create new data.table by keeping only certain columns or rows
}
\examples{
library(data.table)
DT <- data.table(
  id = c(1,2),
  v1 = c(1,1),
  v2 = c(2,1)
)
keep(DT, i = id==1)
keep(DT, id, v2)
keep(DT, -id)
}

