% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set.R
\name{keep}
\alias{keep}
\alias{keep_}
\title{Create new data.table by keeping only certain columns (equivalent to dplyr::select)}
\usage{
keep(x, ...)

keep_(x, vars)
}
\arguments{
\item{x}{a data.table}

\item{...}{Variables to keep. Default to all. See the \link[dplyr]{select} documentation.}

\item{vars}{Used to work around non-standard evaluation.}
}
\description{
Create new data.table by keeping only certain columns (equivalent to dplyr::select)
}
\examples{
library(data.table)
DT <- data.table(
  id = c(1,2),
  v1 = c(1,1),
  v2 = c(2,1)
)
keep(DT, id, v2)
keep(DT, -id)
}

