% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lead-lag}
\alias{lag.default}
\alias{lead}
\alias{lead-lag}
\title{lead and lag.}
\usage{
lead(x, n = 1L, order_by, along_with, units = NULL, default = NA, ...)

\method{lag}{default}(x, n = 1L, order_by, along_with, units = NULL,
  default = NA, ...)
}
\arguments{
\item{x}{a vector of values}

\item{n}{a positive integer of length 1, giving the number of positions to lead or lag by}

\item{order_by}{override the default ordering to use another vector}

\item{along_with}{compute lag with respect to this vector instead of previous row}

\item{units}{computes lag in units of "day", "week", "month", "quarter" or "year" periods. Should be used when \code{along_with} is a date.}

\item{default}{value used for non-existant rows. Defaults to \code{NA}.}
}
\description{
lead and lag are useful for comparing values for date offset by a constant
}
\examples{
# along_with
DT <- data.table(
 id    = c(1, 1, 1, 1, 1, 2, 2),
 date  = c(1992, 1989, 1991, 1990, 1994, 1992, 1991),
 value = c(4.1, 4.5, 3.3, 5.3, 3.0, 3.2, 5.2)
)
DT \%>\% group_by(id) \%>\% mutate(lag(value, 1, order_by = date)) # wrong
DT \%>\% group_by(id) \%>\% mutate(lag(value, 1, along_with = date)) # right
# units
value <- c(4.1, 4.5, 3.3, 5.3, 3.0, 3.2, 5.2)
date <- as.Date(c("03/01/1992", "04/03/1992", "07/15/1992", "08/21/1992"), "\%m/\%d/\%Y")
DT \%>\% group_by(id) \%>\% mutate(lag(value, 1, along_with = date, units = "month"))
}

