% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lead-lag}
\alias{lag}
\alias{lead}
\alias{lead-lag}
\title{Lead and lag.}
\usage{
lead(x, n = 1L, default = NA, order_by = NULL, along_with = NULL, units,
  ...)

lag(x, n = 1L, default = NA, order_by = NULL, along_with = NULL, units,
  ...)
}
\arguments{
\item{x}{a vector of values}

\item{n}{a postive integer of length 1, giving the number of positions to
lead or lag by}

\item{default}{value used for non-existant rows. Defaults to \code{NA}.}

\item{order_by}{override the default ordering to use another vector}

\item{along_with}{instead of using order_by, along_with computes lag/lead variable based on the variable gien in along_with - n}

\item{units}{when variable in along_with is a daily date, computes lag based on "day", "week", "month", "quarter" or "year". Internally, date are matched based on their largest round unit not greater than the corresponding date.}

\item{...}{Needed for compatibility with lag generic.}
}
\description{
Lead and lag are useful for comparing values offset by a constant (e.g. the
previous or next value)
}
\examples{
lead(1:10, 1)
lead(1:10, 2)

lag(1:10, 1)
lead(1:10, 1)

x <- runif(5)
cbind(ahead = lead(x), x, behind = lag(x))

# Use order_by if data not already ordered
df <- data.frame(year = 2000:2005, value = (0:5) ^ 2)
scrambled <- df[sample(nrow(df)), ]
right <- mutate(scrambled, prev = lag(value, order_by = year))
arrange(right, year)
Use along_with if unbalanced panel data
library(lubridate)
date  <-  dmy(c("01031992","03041992","05051992","21081992"))
value <-  c(4.1,4.5,3.3,5.3)
value_lag <- lead(value,1,along_with = date_date, units = "month")
}

