% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lead-lag}
\alias{lag.default}
\alias{lead}
\alias{lead-lag}
\title{lead and lag.}
\usage{
lead(x, n = 1L, order_by = NULL, along_with = NULL, default = NA, ...)

\method{lag}{default}(x, n = 1L, order_by = NULL, along_with = NULL,
  default = NA, ...)
}
\arguments{
\item{x}{a vector of values}

\item{n}{a positive integer of length 1, giving the number of positions to lead or lag by. When the package lubridate is loaded, it can be a period when using with along_with (see the lubridate function minutes, hours, days, weeks, months and years)}

\item{order_by}{override the default ordering to use another vector}

\item{along_with}{compute lag with respect to this vector instead of previous row}

\item{default}{value used for non-existant rows. Defaults to \code{NA}.}
}
\description{
lead and lag are useful for comparing values for date offset by a constant
}
\examples{
# along_with
DT <- data.table(
 id    = c(1, 1, 1, 1, 1, 2, 2),
 date  = c(1992, 1989, 1991, 1990, 1994, 1992, 1991),
 value = c(4.1, 4.5, 3.3, 5.3, 3.0, 3.2, 5.2)
)
DT \%>\% group_by(id) \%>\% mutate(lag(value, 1, order_by = date)) # wrong
DT \%>\% group_by(id) \%>\% mutate(lag(value, 1, along_with = date)) # right
# units
library(lubridate)
df <- data.frame(
    id = c("1", "1", "1", "1"),
  date = mdy(c("03/01/1992", "04/03/1992", "07/15/1992", "08/21/1992"), "\%m/\%d/\%Y"),
 value = c(4.1, 4.5, 3.3, 5.3)
  )
df <- df \%>\% mutate(datem = floor_date(date, "month"))
DT \%>\% group_by(id) \%>\% mutate(lag(value, months(1), along_with = datem))
}

