% Generated by roxygen2 (4.0.2): do not edit by hand
\name{n_unique}
\alias{n_unique}
\title{returns number of unique element}
\usage{
n_unique(x, ...)
}
\arguments{
\item{x}{a vector or a data.table}

\item{...}{generics compatibility}
}
\value{
number of unique element in an object
}
\description{
returns number of unique element
}
\examples{
n_unique(c(1,2,3))
library(data.table)
DT <- data.table(a = rep(1:2, each = 3), b = 1:6)
n_unique(DT)
n_unique(DT, by = "a")
}

