% Generated by roxygen2 (4.0.2): do not edit by hand
\name{roll}
\alias{roll}
\title{Apply rollling functions}
\usage{
roll(x, FUN, n, along_with = NULL, order_by = NULL, ...)
}
\arguments{
\item{x}{a vector}

\item{FUN}{function to apply}

\item{n}{an integer. Window is all the observations in \code{[along_with - n, along_with]}}

\item{along_with}{Integer variable specifying dates. NA are not accepted}

\item{...}{options for the function}
}
\description{
Apply rollling functions
}
\examples{
along_with  = c(1, 2, 4, 7)
x = c(1, 1, 1, 1)
roll(x, sum, n = 1, along_with = along_with)
roll(x, sum, n = 2, along_with = along_with)
roll(x, sum, n = -1, along_with = along_with)
roll(x, sum, n = -2, along_with = along_with)
}

