% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sample_mean}
\alias{sample_mean}
\title{Weighted mean}
\usage{
sample_mean(x, na.rm = TRUE, w = rep(1L, length(x)), min = 0)
}
\arguments{
\item{x}{A vector of values}

\item{w}{vector of weights.}

\item{min}{If length of vector x once missing x and w and zero value of w are removed is strictly lower than min, the returned value is NA.}
}
\value{
When w / min are not specified,  weighted_mean(, min = 0) is the same as mean(., na.rm = TRUE) while weighted_mean(, min = length(x)) is the same as mean(.).
}
\description{
Weighted mean
}

