% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sample_mode}
\alias{sample_mode}
\title{Statistical mode}
\usage{
sample_mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector of values}

\item{na.rm}{Should NA values dropped?}
}
\value{
Returns one mode of the vector (in case of ties, the first value is chosen)
}
\description{
Statistical mode
}
\examples{
sample_mode(c(1, 2, 2))
sample_mode(c(1, 2))
sample_mode(c(NA,NA,1))
sample_mode(c(NA,NA,1), na.rm = TRUE)
}

