% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setcols}
\alias{setcols}
\alias{setcols_}
\title{Retain certain columns of a data.table in place (= Stata keep).}
\usage{
setcols(x, ...)

setcols_(x, ..., .dots)
}
\arguments{
\item{x}{a data.table}

\item{...}{Variables to keep. See the \link[dplyr]{select} documentation.}

\item{.dots}{Used to work around non-standard evaluation.}
}
\description{
Retain certain columns of a data.table in place (= Stata keep).
}
\examples{
library(data.table)
DT <- data.table(
  id = c(1,2),
  v1 = c(1,1),
  v2 = c(2,1)
)
setcols(DT, id, v2)
setcols(DT, -id)
}

