% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sum_up}
\alias{sum_up}
\alias{sum_up_}
\title{Gives summary statistics (corresponds to Stata command summarize)}
\usage{
sum_up(.data, ..., by = NULL, d = FALSE)

sum_up_(.data, ..., .dots, by = NULL, d = FALSE)
}
\arguments{
\item{...}{Variables to include/exclude. Defaults to all non-grouping variables. See the \link[dplyr]{select} documentation.}

\item{by}{Print summary statistics by group. Default to NULL (since it is the most common behavior)}

\item{d}{Should detailed summary statistics be printed?}

\item{DT}{A tbl_dt or tbl_grouped_dt.}
}
\description{
Gives summary statistics (corresponds to Stata command summarize)
}
\examples{
library(data.table)
library(dplyr)
N <- 100; K <- 10
DT <- data.table(
  id = 1:N,
  v1 =  sample(5, N, TRUE),
  v2 =  sample(1e6, N, TRUE),
  v3 =  sample(round(runif(100, max = 100), 4), N, TRUE)
)
sum_up(DT)
sum_up(DT, v2, d=T)
sum_up(DT, starts_with("v"), by = v1)
}

