% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sum_up}
\alias{sum_up}
\alias{sum_up_}
\title{Gives summary statistics (corresponds to Stata command summarize)}
\usage{
sum_up(x, ..., d = FALSE, w = NULL, na.rm = TRUE, by = NULL,
  graph = FALSE, minimal = TRUE)

sum_up_(x, ..., .dots, d = FALSE, w = NULL, na.rm = TRUE, by = NULL,
  graph = FALSE, minimal = TRUE)
}
\arguments{
\item{x}{a data.table}

\item{...}{Variables to include. Defaults to all non-grouping variables. See the \link[dplyr]{select} documentation.}

\item{d}{Should detailed summary statistics be printed?}

\item{w}{Weights. Default to NULL.}

\item{by}{Groups within which summary statistics are printed. Default to NULL. See the \link[dplyr]{select} documentation.}
}
\description{
Gives summary statistics (corresponds to Stata command summarize)
}
\examples{
N <- 1000; K <- 10
DT <- data.table(
  id = 1:N,
  v1 = sample(5, N, TRUE),
  v2 = sample(1e6, N, TRUE)
)
sum_up(DT)
sum_up(DT, v2, d = T)
sum_up(DT, starts_with("v"), by = v1)
sum_up(DT, by = v1, graph = TRUE)
}

