% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_up.R
\name{sum_up}
\alias{sum_up}
\title{Gives summary statistics (deprecated)}
\usage{
sum_up(...)
}
\arguments{
\item{...}{arguments}
}
\value{
a data.frame
}
\description{
Gives summary statistics (deprecated)
}
\examples{
library(dplyr)
N <- 100
df <- data_frame(
  id = 1:N,
  v1 = sample(5, N, TRUE),
  v2 = sample(1e6, N, TRUE)
)
sum_up(df)
sum_up(df, v2, d = TRUE)
sum_up(df, v2, wt = v1)
df \%>\% group_by(v1) \%>\% sum_up(starts_with("v"))
}
