% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tab}
\alias{tab}
\alias{tab_}
\title{Returns cross tabulation}
\usage{
tab(x, ..., i = NULL, w = NULL)

tab_(x, vars = NULL, i = NULL, w = NULL)
}
\arguments{
\item{x}{a data.table}

\item{...}{Variable to include. If length is two, a special cross tabulation table is printed although the a long data.table is always (invisibly) returned.}

\item{i}{Condition to apply function on certain rows only}

\item{w}{Weights. Default to NULL.}

\item{vars}{Used to work around non-standard evaluation.}
}
\value{
a data.table sorted by variables in by, var, and with a new column "N" for counts.
}
\description{
Returns cross tabulation
}
\examples{
library(data.table)
N <- 1e2 ; K = 10
DT <- data.table(
  id = sample(c(NA,1:5), N/K, TRUE),
  v1 =  sample(c(NA,1:5), N, TRUE),
  v2 =  sample(c(NA,1:1e6), N, TRUE)
)
tab(DT, id)
tab(DT, id, v1, w = v2)
}

