% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\alias{tab_}
\title{Returns cross tabulation}
\usage{
tab(x, ...)

tab_(x, vars = NULL, i = NULL, w = NULL, na.omit = TRUE)
}
\arguments{
\item{x}{a vector or a data.table}

\item{...}{Variable to include. If length is two, a special cross tabulation table is printed although the a long data.table is always (invisibly) returned.}

\item{vars}{Used to work around non-standard evaluation.}

\item{i}{Condition to apply function on certain rows only}

\item{w}{Frequency weights. Default to NULL.}
}
\value{
a data.table sorted by variables in ..., and with a columns "Freq", "Percent", and "Cum." for counts.
}
\description{
Returns cross tabulation
}
\examples{
library(data.table)
N <- 1e2 ; K = 10
DT <- data.table(
  id = sample(c(NA,1:5), N/K, TRUE),
  v1 =  sample(c(NA,1:5), N, TRUE),
  v2 =  sample(c(NA,1:1e6), N, TRUE)
)
tab(DT[["id"]])
tab(DT, id)
tab(DT, id, v1, w = v2,)
}

