% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\alias{tab_}
\title{Returns cross tabulation}
\usage{
tab(x, ...)

tab_(x, vars = NULL, i = NULL, w = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{a vector or a data.frame}

\item{...}{Variable to include. If length is two, a special cross tabulation table is printed although the a long data.frame is always (invisibly) returned.}

\item{vars}{Used to work around non-standard evaluation.}

\item{i}{Condition to apply function on certain rows only}

\item{w}{Frequency weights. Default to NULL.}

\item{na.rm}{Remove missing values. Default to FALSE}
}
\value{
a data.frame sorted by variables in ..., and with a columns "Freq", "Percent", and "Cum." for counts.
}
\description{
Returns cross tabulation
}
\examples{
library(dplyr)
N <- 1e2 ; K = 10
df <- data.frame(
  id = sample(c(NA,1:5), N/K, TRUE),
  v1 =  sample(c(NA,1:5), N/K, TRUE)
)
tab(df[["id"]])
tab(df, id)
df \%>\% group_by(id) \%>\% tab()
df \%>\% group_by(id) \%>\% tab(v1)
tab(df, id, i = id>=3)
}

