% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tag}
\alias{tag}
\title{Tag a numeric vector (= Stata tag)}
\usage{
tag(x, fromLast = FALSE)
}
\arguments{
\item{x}{A vector of values}
}
\value{
A vector with only 1
}
\description{
Tag a numeric vector (= Stata tag)
}
\examples{
tag(c(1,2))
tag(c(1,2), fromLast = TRUE)
}

