% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tag}
\alias{tag}
\title{Creates a vector of zero except for one subscript}
\usage{
tag(n, fromLast = FALSE)
}
\arguments{
\item{n}{An integer (size of output)}

\item{fromLast}{Should 1 be at the end?}
}
\value{
A vector of zeros except one subscript equal to one.
}
\description{
Creates a vector of zero except for one subscript
}
\examples{
tag(2)
tag(2, fromLast = TRUE)
}

