% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tempname}
\alias{tempname}
\title{Create unique names within environment}
\usage{
tempname(where = globalenv(), prefix = "temp", inherits = TRUE)
}
\arguments{
\item{where}{A chracter vector, list or an environment}

\item{prefix}{A character vector that specifies prefix for new name}

\item{inherits}{Should the name unique also in the enclosing frames of the environment?}

\item{n}{Number of names to generate}
}
\description{
Create unique names within environment
}
\examples{
tempname(c("temp","temp1"))
tempname(globalenv())
tempname(globalenv())
tempname(data.table(temp= 1))
}

