% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tempname}
\alias{tempname}
\title{Create unique names within environment}
\usage{
tempname(prefix, where, inherits = TRUE)
}
\arguments{
\item{prefix}{A character that specifies prefix for new name}

\item{where}{A list or an environment}

\item{inherits}{Should the name unique also in the enclosing frames of the environment?}
}
\description{
Create unique names within environment
}
\examples{
temp <- 1
tempname <- tempname("temp", globalenv())
DT <- data.table(
 temp = c(1, 1, 1, 1, 1, 2, 2),
 temp1 = c(1992, 1989, 1991, 1990, 1994, 1992, 1991),
)
tempvar <- tempname("temp", DT)
tempvar <- tempname("temporary", DT)
}

