% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempname.R
\name{tempname}
\alias{tempname}
\title{Create unique names within a list, a data.frame, or an environment}
\usage{
tempname(where = globalenv(), n = 1, prefix = ".temp", inherits = TRUE)
}
\arguments{
\item{where}{A chracter vector, list or an environment}

\item{n}{An integar that specifies length of the output}

\item{prefix}{A character vector that specifies prefix for new name}

\item{inherits}{Should the name unique also in the enclosing frames of the environment?}
}
\description{
Create unique names within a list, a data.frame, or an environment
}
\examples{
tempname(c("temp1", "temp3"), 4)
tempname(globalenv())
tempname(data.frame(temp = 1), n = 3)
}
