% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vars}
\alias{vars}
\title{select variables in a data.frame}
\usage{
vars(x, ...)
}
\arguments{
\item{x}{a data.frame}

\item{...}{}
}
\value{
a character vector. All the syntax available in \code{select_vars} are allowed. Characters are understood as wildcards.
}
\description{
select variables in a data.frame
}
\examples{
df <- data.frame(id = 3, id2 = 4)
vars(df, id)
vars(df, "id*")
}

