% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vars.R
\name{vars}
\alias{vars}
\alias{vars_}
\alias{vars_not}
\alias{vars_not_}
\title{select variables in a data.frame}
\usage{
vars(x, ...)

vars_not(x, ...)

vars_not_(x, .dots)

vars_(x, .dots)
}
\arguments{
\item{x}{a data.frame}

\item{...}{See the \link[dplyr]{select} documentation.}

\item{.dots}{Needed to work around non-standard evaluation}
}
\value{
a character vector. All the syntax available in \code{select_vars} are allowed. Characters are understood as wildcards.
}
\description{
select variables in a data.frame
}
\examples{
df <- data.frame(id = 3, id2 = 4)
vars(df, id)
vars(df, w("id*"))
}

